+++ Services/TimeTable.php
@@ -1,14 +1,20 @@
 <?php
 
 namespace Leantime\Plugins\TimeTable\Services;
+
 use Leantime\Plugins\TimeTable\Repositories\TimeTable as TimeTableRepository;
+use Carbon\CarbonImmutable;
 
-
+/**
+ * Time table services file.
+ */
 class TimeTable
 {
-
     private TimeTableRepository $timeTableRepo;
 
+    /**
+     * @var array<string, string>
+     */
     private static array $assets = [
         // source => target
         __DIR__ . '/../assets/timeTable.js' => APP_ROOT . '/public/dist/js/plugin-timeTable.v%%VERSION%%.js',
@@ -18,12 +24,19 @@ class TimeTable
     /**
      * constructor
      *
-     * @param TimeTableRepository $timeTableRepo
+     * @param  TimeTableRepository $timeTableRepo
      * @return void
      */
-    public function __construct(TimeTableRepository $timeTableRepo) {
-      $this->timeTableRepo = $timeTableRepo;
+    public function __construct(TimeTableRepository $timeTableRepo)
+    {
+        $this->timeTableRepo = $timeTableRepo;
     }
+
+    /**
+     * Install plugin.
+     *
+     * @return void
+     */
     public function install(): void
     {
         foreach (self::getAssets() as $source => $target) {
@@ -34,6 +47,11 @@ class TimeTable
         }
     }
 
+    /**
+     * Uninstall plugin.
+     *
+     * @return void
+     */
     public function uninstall(): void
     {
         foreach (self::getAssets() as $target) {
@@ -51,5 +69,41 @@ class TimeTable
     private static function getAssets(): array
     {
         return self::$assets;
+    }
+
+    /**
+     * @return array<array<string, string>>
+     */
+    public function getUniqueTicketIds(CarbonImmutable $dateFrom, CarbonImmutable $dateTo): array
+    {
+        return $this->timeTableRepo->getUniqueTicketIds($dateFrom, $dateTo);
+    }
+
+    /**
+     * @return array<array<string, string>>
+     */
+    public function getTimesheetByTicketIdAndWorkDate(string $ticketId, CarbonImmutable $workDate, ?string $searchTerm): array
+    {
+        return $this->timeTableRepo->getTimesheetByTicketIdAndWorkDate($ticketId, $workDate, $searchTerm);
+    }
+
+     /**
+     * updateTime - update specific time entry
+     *
+     * @return void
+     */
+    public function logTimeOnTicket(array $values): void
+    {
+        $this->timeTableRepo->logTimeOnTicket($values);
+    }
+
+    /**
+     * updateTime - update specific time entry
+     *
+     * @return void
+     */
+    public function updateTime(array $values): void
+    {
+        $this->timeTableRepo->updateTime($values);
     }
 }